backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-oauth-toolkit==1.7.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
Pillow==10.1.0
django-storages==1.14.2
boto3==1.34.0
django-environ==0.11.2
django-ratelimit==4.1.0
django-extensions==3.2.3
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==23.5
drf-spectacular==0.27.0
cryptography
django-two-factor-auth==1.15.0
qrcode==7.4.2
django-axes==6.1.1
django-csp==3.7
django-permissions-policy==4.17.0
django-security
django-audit-log
sentry-sdk==1.38.0
